Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/branches/NET_2_0@636541 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/net/ntp/NtpV3Impl.java
@@ -128,7 +128,7 @@ public class NtpV3Impl implements NtpV3Packet
      */
     public int getPoll()
     {
-        return (int) (buf[POLL_INDEX]);
+        return buf[POLL_INDEX];
     }
 
     /***
@@ -150,7 +150,7 @@ public class NtpV3Impl implements NtpV3Packet
      */
     public int getPrecision()
     {
-        return (int) buf[PRECISION_INDEX];
+        return buf[PRECISION_INDEX];
     }
 
     /***
